add transferParts to partsController

+++ controller/PartController.js
@@ -19,7 +19,7 @@ export const createPart = async (req, res, next) => {
 
     await newPart.save();
 
-    const parts = await Part.find({userId:  userData.user._id})
+    const parts = await Part.find({ userId: userData.user._id });
 
     res.status(200).json({ newPart: parts, message: "Part created" });
   } catch (error) {
@@ -29,15 +29,8 @@ export const createPart = async (req, res, next) => {
 
 export const editPart = async (req, res, next) => {
   try {
-    const {
-      _id,
-      partNumber,
-      name,
-      quantity,
-      price,
-      lowLimit,
-      description,
-    } = req.body;
+    const { _id, partNumber, name, quantity, price, lowLimit, description } =
+      req.body;
 
     const userData = await dataFunction(req, res, next);
 
@@ -69,7 +62,7 @@ export const editPart = async (req, res, next) => {
       {
         new: true,
         runValidators: true,
-      }
+      },
     );
 
     if (!updatedPart) {
@@ -91,7 +84,6 @@ export const editPart = async (req, res, next) => {
   }
 };
 
-
 export const getData = async (req, res, next) => {
   try {
     const parts = await Part.find({ userId: req.user.id, deleted: false });
@@ -105,10 +97,9 @@ export const softDelete = async (req, res, next) => {
   try {
     const userData = await dataFunction(req, res, next);
 
-    const deletePart = await Part.findByIdAndUpdate(
-      req.body._id,
-      { deleted: true }
-    );
+    const deletePart = await Part.findByIdAndUpdate(req.body._id, {
+      deleted: true,
+    });
 
     if (!deletePart) {
       return res.status(404).json({
@@ -118,7 +109,7 @@ export const softDelete = async (req, res, next) => {
 
     await deletePart.save();
 
-    const parts = await Part.find({userId:  userData.user._id})
+    const parts = await Part.find({ userId: userData.user._id });
 
     res.status(200).json({
       message: "Part deleted!",
@@ -130,18 +121,92 @@ export const softDelete = async (req, res, next) => {
 };
 
 export const hardDelete = async (req, res, next) => {
-  console.log("test");
   try {
+    const userData = await dataFunction(req, res, next);
 
-   const userData = await dataFunction(req, res, next);
+    await Part.deleteMany({
+      userId: userData.user._id,
+    });
 
-await Part.deleteMany({
-  userId: userData.user._id,
-});
+    res.status(200).json({
+      message: "All parts deleted!",
+    });
+  } catch (error) {
+    next(error);
+  }
+};
 
-res.status(200).json({
-  message: "All parts deleted!",
-});
+
+export const transferData = async (req, res, next) => {
+  try {
+    const {
+      sourceEmail,
+      sourcePassword,
+      targetEmail,
+      targetPassword,
+      deleteAfterTransfer,
+    } = req.body;
+
+    const sourceUser = await User.findOne({
+      email: sourceEmail,
+    });
+
+    if (!sourceUser) {
+      return res.status(404).json({
+        message: "Quellbenutzer nicht gefunden",
+      });
+    }
+
+    const sourcePasswordCorrect = await bcrypt.compare(
+      sourcePassword,
+      sourceUser.password,
+    );
+
+    if (!sourcePasswordCorrect) {
+      return res.status(401).json({
+        message: "Passwort des Quellbenutzers falsch",
+      });
+    }
+
+    const targetUser = await User.findOne({
+      email: targetEmail,
+    });
+
+    if (!targetUser) {
+      return res.status(404).json({
+        message: "Zielbenutzer nicht gefunden",
+      });
+    }
+
+    const targetPasswordCorrect = await bcrypt.compare(
+      targetPassword,
+      targetUser.password,
+    );
+
+    if (!targetPasswordCorrect) {
+      return res.status(401).json({
+        message: "Passwort des Zielbenutzers falsch",
+      });
+    }
+
+    if (deleteAfterTransfer) {
+      await Part.deleteMany({
+        userId: sourceUser._id,
+      });
+    } else {
+      await Part.updateMany(
+        {
+          userId: sourceUser._id,
+        },
+        {
+          userId: targetUser._id,
+        },
+      );
+    }
+
+    res.status(200).json({
+      message: "Transfer erfolgreich",
+    });
   } catch (error) {
     next(error);
   }
